chore: move util functions to lib.util
