+++ intelmq_webinput_csv/lib/util.py
@@ -48,3 +48,64 @@ def handle_extra(value: str) -> dict:
         if not isinstance(value, dict):
             value = {'data': value}
     return value
+
+def write_temp_file(data):
+    """
+    Write metadata about the current active file.
+    filename, total_lines
+    """
+    with open(TEMP_FILE, 'wb') as handle:
+        pickle.dump(data, handle)
+
+
+def get_temp_file():
+    """
+    Opposite of write_temp_file
+    """
+    try:
+        with open(TEMP_FILE, 'rb') as handle:
+            data = pickle.load(handle)
+            if len(data) == 2:
+                return data
+    except TypeError:  # TypeError: returned value has no len()
+        return False
+    return False
+
+
+def handle_parameters(form):
+    parameters = {}
+    for key, default_value in CONFIG.items():
+        parameters[key] = form.get(key, default_value)
+    for key, value in PARAMETERS.items():
+        parameters[key] = form.get(key, value)
+    parameters['dryrun'] = json.loads(parameters['dryrun'])
+    if parameters['dryrun']:
+        parameters['classification.type'] = 'test'
+        parameters['classification.identifier'] = 'test'
+    if type(parameters['columns']) is not list and parameters['use_column']:
+        parameters['use_column'] = [json.loads(a.lower()) for a in
+                                    parameters['use_column'].split(',')]
+        parameters['columns'] = parameters['columns'].split(',')
+    parameters['columns'] = [a if b else None for a, b in
+                             zip(parameters['columns'],
+                                 parameters['use_column'])]
+    parameters['skipInitialLines'] = int(parameters['skipInitialLines'])
+    parameters['skipInitialSpace'] = json.loads(parameters['skipInitialSpace'])
+    parameters['has_header'] = json.loads(parameters['has_header'])
+    parameters['loadLinesMax'] = int(parameters['loadLinesMax'])
+    return parameters
+
+
+def create_response(text, content_type=None):
+    is_json = False
+    if not isinstance(text, str):
+        text = jsonify(text)
+        is_json = True
+    response = make_response(text)
+    if is_json:
+        response.mimetype = 'application/json'
+        response.headers['Content-Type'] = "text/json; charset=utf-8"
+    if content_type:
+        response.headers['Content-Type'] = content_type
+    response.headers['Access-Control-Allow-Origin'] = "*"
+    return response
